@@ -18,7 +18,7 @@
 		"@sveltejs/adapter-node": "^5.2.11",
 		"@sveltejs/adapter-static": "^3.0.8",
 		"@sveltejs/kit": "^2.15.1",
-		"@sveltejs/vite-plugin-svelte": "^3.1.2",
+		"@sveltejs/vite-plugin-svelte": "^4.0.4",
 		"@types/eslint": "^8.56.12",
 		"eslint": "^9.17.0",
 		"eslint-config-prettier": "^9.1.0",
@@ -26,8 +26,8 @@
 		"globals": "^15.14.0",
 		"prettier": "^3.4.2",
 		"prettier-plugin-svelte": "^3.3.2",
-		"svelte": "^4.2.19",
-		"svelte-check": "^3.8.6",
+		"svelte": "^5.16.0",
+		"svelte-check": "^4.1.1",
 		"tslib": "^2.8.1",
 		"typescript": "^5.7.2",
 		"typescript-eslint": "8.0.0-alpha.47",
